Rename pages -> web_pages, create package web_tests
