+++ src/main/java/wbd/web_pages/LoginPage.java
@@ -1,4 +1,4 @@
-package wbd.pages;
+package wbd.web_pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
